web/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-rest-swagger==2.2.0
django-taggit==1.3.0
django-unixdatetimefield==1.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
fonttools==4.38.0
greenlet==2.0.1
h11==0.14.0
hiredis==2.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.3
msgpack==1.0.4
msgpack-python==0.5.6
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.2
pandas-datareader==0.10.0
pandas-ta==0.3.14b0
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.2
rel==0.4.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
style==1.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.4.2
zope.interface==5.5.2
